Extract travel composer catalog runtime
